feat(ng-add): update behavior and code cleanup
