+++ packages/semver/src/schematics/ng-add/index.ts
@@ -2,7 +2,7 @@ import { chain, Rule, Tree } from '@angular-devkit/schematics';
 import { updateNxJsonInTree, updateWorkspace } from '@nrwl/workspace';
 
 import { SchemaOptions } from './schema';
-import { addDependencies } from './utils/dependencies';
+import { addDependencies } from './utils/add-dependencies';
 import { updateWorkspaceFromPrompt, updateWorkspaceFromSchema } from './utils/workspace';
 
 export function ngAdd(options: SchemaOptions): Rule {
